# api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: CRM System API
  description: |
    Complete API documentation for the Tour Operator CRM system.

    This API provides endpoints for managing:
    - Suppliers (Hotels, Guides, Vehicles, Tour Packages, Restaurants, Transfers, Extra Expenses, Entrance Fees)
    - Quotations and Itineraries
    - Customer Requests
    - Invoices (Receivable & Payable)
    - Finance and Business Metrics
    - Providers/Companies
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-production-domain.com
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    Page:
      in: query
      name: page
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Sort:
      in: query
      name: sort
      description: Comma-separated list of fields to sort by. Prefix with '-' for descending order.
      schema:
        type: string
        example: "-created_at,name"
    Search:
      in: query
      name: search
      description: Search query string
      schema:
        type: string
    StatusFilter:
      in: query
      name: status
      description: Filter by status
      schema:
        type: string
        enum: [all, active, inactive]
...
  schemas:
...
    PageMeta:
      type: object
      description: Pagination metadata
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 25
        total:
          type: integer
          description: Total number of items
          example: 100
    PagedResponse:
      type: object
      description: Generic paged response wrapper
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          format: int64
        organization_id:
          type: integer
          format: int64
        provider_id:
          type: integer
          format: int64
          nullable: true
        provider_name:
          type: string
          nullable: true
        vehicle_type:
          type: string
        max_capacity:
          type: integer
        city:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/vehicles:
    get:
      tags:
        - Suppliers
      summary: List all vehicles
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/StatusFilter'
        - name: city
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Vehicle'
    post:
      tags:
        - Suppliers
      summary: Create new vehicle
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            $ref: '#/components/headers/IdempotencyKey/schema'
      responses:
        '201':
          description: Vehicle created successfully
          headers:
            Location:
              schema:
                type: string
  /api/vehicles/{id}:
    get:
      tags:
        - Suppliers
      summary: Get vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    patch:
      tags:
        - Suppliers
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    delete:
      tags:
        - Suppliers
      summary: Archive vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Vehicle archived successfully
...